Move induction list components into a shared directory.
